Move to executor_utils::validateIndexCasts
